backend/requirements.txt
anyio==3.6.1
asgiref==3.5.2
beautifultable==0.8.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
Django==3.2
django-colorfield==0.7.2
django-cors-headers==3.13.0
djangorestframework==3.13.1
fandogh-cli==1.26.10
fastapi==0.79.0
idna==3.3
Pillow==9.2.0
pycparser==2.21
pydantic==1.9.1
pyOpenSSL==22.0.0
pytz==2022.1
PyYAML==5.4.1
requests==2.28.1
requests-toolbelt==0.9.1
retrying==1.3.3
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.19.1
typing_extensions==4.3.0
tzlocal==2.1
urllib3==1.26.11
websocket-client==1.3.3
